Revert "Consolidate on-prem-installer deb packages (#363)"
This reverts commit 2f5893046a9eb612fcda0cde535f2e42ad2d3622.
