+++ on-prem-installers/cmd/onprem-argo-cd/after-remove.sh
@@ -6,13 +6,6 @@
 
 set -o errexit
 
-export KUBECONFIG=/home/$USER/.kube/config
-
-# Add /usr/local/bin to the PATH as some utilities, like kubectl, could be installed there
-export PATH=$PATH:/usr/local/bin
-
-remove_argocd() {   
-
 cat << "EOF"
 
      _                     ____ ____    ____
@@ -24,37 +17,18 @@ cat << "EOF"
 
 EOF
 
-    # If ArgoCD is upgraded its helm chart shouldn't be deleted, as helm upgrade
-    # will be called in after-upgrade
-    if [ "${1}" = "upgrade" ]; then
-        return 0
-    fi
+# If ArgoCD is upgraded its helm chart shouldn't be deleted, as helm upgrade
+# will be called in after-upgrade
+if [ "${1}" = "upgrade" ]; then
+    exit 0
+fi
 
-    helm delete argocd -n argocd || true
+export KUBECONFIG=/home/$USER/.kube/config
 
-    # Remove artifacts
-    rm -rf /tmp/argo-cd || true
-}
+# Add /usr/local/bin to the PATH as some utilities, like kubectl, could be installed there
+export PATH=$PATH:/usr/local/bin
 
+helm delete argocd -n argocd || true
 
-
-remove_gitea() {
-cat << "EOF"
-
-   ____ _ _               ____
-  / ___(_) |_ ___  __ _  |  _ \ ___ _ __ ___   _____   _____
- | |  _| | __/ _ \/ _` | | |_) / _ \ '_ ` _ \ / _ \ \ / / _ \
- | |_| | | ||  __/ (_| | |  _ <  __/ | | | | | (_) \ V /  __/
-  \____|_|\__\___|\__,_| |_| \_\___|_| |_| |_|\___/ \_/ \___|
-
-
-EOF
-
-    helm delete gitea -n gitea || true
-    kubectl delete secret gitea-cred gitea-tls-certs gitea-token -n gitea || true
-    # clean the certificate on the system
-    rm -f /usr/local/share/ca-certificates/gitea_cert.crt || true
-}
-
-remove_gitea
-remove_argocd
+# Remove artifacts
+rm -rf /tmp/argo-cd || true
