+++ on-prem-installers/cmd/onprem-config-installer/after-install.sh
@@ -1,0 +1,13 @@
+#!/usr/bin/env bash
+
+# SPDX-FileCopyrightText: 2025 Intel Corporation
+#
+# SPDX-License-Identifier: Apache-2.0
+
+set -o errexit
+
+# Add /usr/local/bin to the PATH as some utilities, like kubectl, could be installed there
+export PATH=$PATH:/usr/local/bin
+
+# Execute the installer with the current directory as context
+/usr/bin/onprem-config-installer
