+++ on-prem-installers/cmd/onprem-config-installer/onprem-config-installer.1
@@ -1,0 +1,18 @@
+\SPDX-FileCopyrightText: 2025 Intel Corporation
+\
+\SPDX-License-Identifier: Apache-2.0
+
+.TH INSTALLER "17" "November 2023" "onprem-config-installer 0.1.0" "User Commands"
+.SH NAME
+onprem-config-installer \- manual page for onprem-config-installer 0.1.0
+.SH DESCRIPTION
+.IP
+USAGE: onprem-config-installer
+.SH "SEE ALSO"
+.IP
+Website: https://github.com/open-edge-platform/edge-manageability-framework/on-prem-installers
+.SH "OTHER"
+.IP
+Made by Intel with ❤️
+.IP
+This program is distributed under Apache 2.0 license.
