+++ on-prem-installers/mage/Magefile.go
@@ -121,6 +121,8 @@ func (b Build) All(ctx context.Context) error {
 	mg.CtxDeps(
 		ctx,
 		b.OnpremKEInstaller,
+		b.OSConfig,
+		b.GiteaInstaller,
 		b.ArgocdInstaller,
 		b.OnPremOrchInstaller,
 	)
@@ -137,6 +139,28 @@ func (b Build) OnpremKEInstaller(ctx context.Context) error {
 	)
 
 	return b.onpremKeInstaller()
+}
+
+// Build the OS-Config Installer package.
+func (b Build) OSConfig(ctx context.Context) error {
+	mg.CtxDeps(
+		ctx,
+		b.Deps,
+		mage.Deps.FPM,
+	)
+
+	return b.osConfigInstaller()
+}
+
+// Build the Gitea Installer package.
+func (b Build) GiteaInstaller(ctx context.Context) error {
+	mg.CtxDeps(
+		ctx,
+		b.Deps,
+		mage.Deps.FPM,
+	)
+
+	return b.giteaInstaller()
 }
 
 // Build the Argo-Cd Installer package.
