+++ on-prem-installers/onprem/onprem_installer.sh
@@ -69,8 +69,10 @@ orch_namespace_list=(
 # Variables that depend on the above and might require updating later, are placed in here
 set_artifacts_version() {
   installer_list=(
+    "onprem-config-installer:${DEPLOY_VERSION}"
     "onprem-ke-installer:${DEPLOY_VERSION}"
     "onprem-argocd-installer:${DEPLOY_VERSION}"
+    "onprem-gitea-installer:${DEPLOY_VERSION}"
     "onprem-orch-installer:${DEPLOY_VERSION}"
   )
 
@@ -604,8 +606,12 @@ mv -f "$repo_file" "$cwd/$git_arch_name/$repo_file"
 cd "$cwd"
 rm -rf "$tmp_dir"
 
+# Run OS Configuration installer
+echo "Installing the OS level configuration..."
+eval "sudo NEEDRESTART_MODE=a DEBIAN_FRONTEND=noninteractive apt-get install -y $cwd/$deb_dir_name/onprem-config-installer_*_amd64.deb"
+echo "OS level configuration installed"
 
-# Run OS Configuration installer and K8s Installer
+# Run K8s Installer
 echo "Installing RKE2..."
 if [[ -n "${DOCKER_USERNAME}" && -n "${DOCKER_PASSWORD}" ]]; then
   echo "Docker credentials provided. Installing RKE2 with Docker credentials"
@@ -613,7 +619,7 @@ if [[ -n "${DOCKER_USERNAME}" && -n "${DOCKER_PASSWORD}" ]]; then
 else
   sudo NEEDRESTART_MODE=a DEBIAN_FRONTEND=noninteractive apt-get install -y "$cwd"/$deb_dir_name/onprem-ke-installer_*_amd64.deb
 fi
-echo "OS level configuration installed and RKE2 Installed"
+echo "RKE2 Installed"
 
 mkdir -p /home/"$USER"/.kube
 sudo cp  /etc/rancher/rke2/rke2.yaml /home/"$USER"/.kube/config
@@ -621,14 +627,17 @@ sudo chown -R "$USER":"$USER"  /home/"$USER"/.kube
 sudo chmod 600 /home/"$USER"/.kube/config
 export KUBECONFIG=/home/$USER/.kube/config
 
-
-# Run argo CD installer
-echo "Installing Gitea & ArgoCD..."
-eval "sudo NEEDRESTART_MODE=a DEBIAN_FRONTEND=noninteractive apt-get install -y $cwd/$deb_dir_name/onprem-argocd-installer_*_amd64.deb"
+# Run gitea installer
+echo "Installing Gitea"
+eval "sudo IMAGE_REGISTRY=${GITEA_IMAGE_REGISTRY} NEEDRESTART_MODE=a DEBIAN_FRONTEND=noninteractive apt-get install -y $cwd/$deb_dir_name/onprem-gitea-installer_*_amd64.deb"
 wait_for_namespace_creation $gitea_ns
 sleep 30s
 wait_for_pods_running $gitea_ns
 echo "Gitea Installed"
+
+# Run argo CD installer
+echo "Installing ArgoCD..."
+eval "sudo NEEDRESTART_MODE=a DEBIAN_FRONTEND=noninteractive apt-get install -y $cwd/$deb_dir_name/onprem-argocd-installer_*_amd64.deb"
 wait_for_namespace_creation $argo_cd_ns
 sleep 30s
 wait_for_pods_running $argo_cd_ns
